Moved executable commands to their own namespace
