+++ erukar/engine/commands/ActionCommand.py
@@ -1,4 +1,4 @@
-from .Command import Command
+from erukar.engine.model.Command import Command
 
 class ActionCommand(Command):
     '''Subclass that defines that the command must be used as an action'''
