+++ erukar/engine/commands/DirectionalCommand.py
@@ -1,4 +1,4 @@
-from erukar.engine.model.ActionCommand import ActionCommand
+from erukar.engine.commands.ActionCommand import ActionCommand
 from erukar.engine.model.Direction import Direction
 
 class DirectionalCommand(ActionCommand):
