--- erukar/engine/commands/Give.py
@@ -1,5 +1,0 @@
-from erukar.engine.model.ActionCommand import ActionCommand
-
-
-class Give(ActionCommand):
-    pass
