--- erukar/engine/commands/Join.py
@@ -1,5 +1,0 @@
-from erukar.engine.model.ActionCommand import ActionCommand
-
-
-class Join(ActionCommand):
-    pass
