--- erukar/engine/commands/Quit.py
@@ -1,5 +1,0 @@
-from erukar.engine.model.ActionCommand import ActionCommand
-
-
-class Quit(ActionCommand):
-    pass
