--- erukar/engine/commands/Use.py
@@ -1,5 +1,0 @@
-from erukar.engine.model.ActionCommand import ActionCommand
-
-
-class Use(ActionCommand):
-    pass
