+++ erukar/engine/commands/__init__.py
@@ -1,35 +1,5 @@
-from erukar.engine.commands.Attack import Attack
-from erukar.engine.commands.Close import Close
-from erukar.engine.commands.DirectionalCommand import DirectionalCommand
-from erukar.engine.commands.Equip import Equip
-from erukar.engine.commands.Give import Give
-from erukar.engine.commands.Inspect import Inspect
-from erukar.engine.commands.Inventory import Inventory
-from erukar.engine.commands.Join import Join
-from erukar.engine.commands.Map import Map
-from erukar.engine.commands.Move import Move
-from erukar.engine.commands.Open import Open
-from erukar.engine.commands.Quit import Quit
-from erukar.engine.commands.Stats import Stats
-from erukar.engine.commands.Take import Take
-from erukar.engine.commands.Unequip import Unequip
-from erukar.engine.commands.Use import Use
+from .executable import *
+from .ActionCommand import ActionCommand
+from .DirectionalCommand import DirectionalCommand
 
-__all__ = [
-    "Attack",
-    "Close",
-    "DirectionalCommand",
-    "Equip",
-    "Give",
-    "Inspect",
-    "Inventory",
-    "Join",
-    "Map",
-    "Move",
-    "Open",
-    "Quit",
-    "Stats",
-    "Take",
-    "Unequip",
-    "Use"
-]
+__all__ = ["ActionCommand", "DirectionalCommand"]
