+++ erukar/engine/commands/executable/Give.py
@@ -1,0 +1,5 @@
+from erukar.engine.commands.ActionCommand import ActionCommand
+
+
+class Give(ActionCommand):
+    pass
