+++ erukar/engine/commands/executable/Inspect.py
