+++ erukar/engine/commands/executable/Join.py
@@ -1,0 +1,5 @@
+from erukar.engine.commands.ActionCommand import ActionCommand
+
+
+class Join(ActionCommand):
+    pass
