+++ erukar/engine/commands/executable/Quit.py
@@ -1,0 +1,5 @@
+from erukar.engine.commands.ActionCommand import ActionCommand
+
+
+class Quit(ActionCommand):
+    pass
