+++ erukar/engine/commands/executable/Use.py
@@ -1,0 +1,5 @@
+from erukar.engine.commands.ActionCommand import ActionCommand
+
+
+class Use(ActionCommand):
+    pass
