+++ erukar/engine/commands/executable/__init__.py
@@ -1,0 +1,33 @@
+from .Attack import Attack
+from .Close import Close
+from .Equip import Equip
+from .Give import Give
+from .Inspect import Inspect
+from .Inventory import Inventory
+from .Join import Join
+from .Map import Map
+from .Move import Move
+from .Open import Open
+from .Quit import Quit
+from .Stats import Stats
+from .Take import Take
+from .Unequip import Unequip
+from .Use import Use
+
+__all__ = [
+    "Attack",
+    "Close",
+    "Equip",
+    "Give",
+    "Inspect",
+    "Inventory",
+    "Join",
+    "Map",
+    "Move",
+    "Open",
+    "Quit",
+    "Stats",
+    "Take",
+    "Unequip",
+    "Use"
+]
